Fix/camelize for structs (#15)
## 🚪 Why?

This Pull Request solves an issue where struct keys were not being
properly converted to camelCase. This caused inconsistencies when
interacting with Mastra API expecting camelCase keys.

This made that the telemetry settings where we need to ensure full
privacy for the inputs and outputs, where not working.

## 🔑 What?

This PR updates the code responsible for serializing structs so that
their keys are correctly converted to camelCase format.