Feat/sort dependencies nullable and reasoning backfilling (#26)
## 🚪 Why?

Zod schemas of type nullable were returned after creating a workflow as
Any or Undefined for Sorbet, which wasn't the intended behavior, we
expected nilable.

Also, the order was not being considered, so dependencies and dependent
structs were not working properly.

Last, we missed to add the backfilling for the reasoning_details that
was covered by Mastra chat, with this we aim to solve this too.

## 🔑 What?

Refactored mastra client to handle reasoning_details, and the
schema_struct_string to handle the sorting of dependencies, and having
nullables translated the right way.

Updated version to the 0.4.2.